Optimize changing directory by making sure the modelReset is only done once we got the proper directory.
Looks as the QFileSystemModel::setDirectory logic is flawn by design cause
it emits rootPathChanged before returning the new QModelIndex for that
root-path. That means that allow following operations done on the
QModelIndex still operate on the previous directory rather then the
new root-path...

In our case that resulted in us reloading the whole directory and
building up the items for the old directory when switching to a
new one. Hallelujah I would say...

So, this commit doubles in fact the speed when changing directories :)